# configs/swagger.yaml
openapi: 3.0.0
info:
  title: NestJS - Open API 3.0.0
  version: 2.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /trademarks:
    get:
      tags:
        - Home
      summary: trademarks
      description: Get all data in HomePage.
      responses:
        200:
          description: Get all Home success
        400:
          description: Invalid status value
...
  /teams:
    get:
      tags:
        - Home
      summary: read teams
      description: read info Homepage.
      parameters:
        - name: page
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: limit
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        200:
          description: Get all Sharing success
        400:
          description: Invalid status value
...
  /users:
    post:
      tags:
        - User
      summary: create new user
      description: Create New
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Create User Successfully
  /users/{id}:
    get:
      tags:
        - User
      summary: read my self
      parameters:
        - name: id
          in: path
          description: Id of the user to get info
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get Info Successfully
        404:
          description: Not Found User
      security:
        - Bearer: []

    put:
      tags:
        - User
      summary: update my self
      description: update user
      operationId: update
      parameters:
        - name: id
          in: path
          description: Id of User
          required: true
          schema:
            type: string
      requestBody:
        description: Update user
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        201:
          description: Updated successfully.
        400:
          description: 1. First name is required;\n2. Last name is required;\n3. Invalid username;\n4.  is not language!
        404:
          description: User not found!|Information not found!
        422:
          description: Create Error
        500:
          description: Internal Server Error
  /admin/users:
    get:
      tags:
        - ManageUser
      summary: get all users
      responses:
        200:
          description: get all successfully.
        404:
          description: User not found!|Information not found!
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    post:
      tags:
        - ManageUser
      summary: create new user
      description: Create New
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Create User Successfully
      security:
        - Bearer: []
  /admin/users/{id}:
    put:
      tags:
        - ManageUser
      summary: update user
      description: update user
      operationId: update
      parameters:
        - name: id
          in: path
          description: Id of User
          required: true
          schema:
            type: string
      requestBody:
        description: Update user
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        201:
          description: Updated successfully.
        400:
          description: 1. First name is required;\n2. Last name is required;\n3. Invalid username;\n4.  is not language!
        404:
          description: User not found!|Information not found!
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - ManageUser
      summary: delete user
      description: Delete User By Id User
      parameters:
        - name: id
          in: path
          description: Id of the User to delete
          required: true
          schema:
            type: string
      responses:
        201:
          description: Delete item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
  /admin/trademarks:
    get:
      tags:
        - Trademark
      summary: trademarks
      description: Get all trademark.
      responses:
        200:
          description: Get all Trademark success
        400:
          description: Invalid status value
      security:
        - Bearer: []
  /admin/trademarks/logo:
    patch:
      tags:
        - Trademark
      summary: update logo trademark
      description: This can only be done by the logged in user.
      parameters:
        - name: type
          in: query
          description: type is (favicon', 'logo-icon', 'logo-word', 'logo') to return
          required: true
          schema:
            type: string
            enum:
              - favicon
              - logo-icon
              - logo-word
              - logo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        201:
          description: Update item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/trademarks/license/{id}:
    patch:
      tags:
        - Trademark
      summary: update license
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: id of license
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: L I E N C E S G O D  2023
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        201:
          description: Update item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - Trademark
      summary: delete license
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: id of license to return
          required: true
          schema:
            type: string
      responses:
        201:
          description: Update item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/trademarks/license:
    post:
      tags:
        - Trademark
      summary: create license
      description: This can only be done by the logged in user.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: L I E N C E S G O D  2023
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        201:
          description: Update item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
  /admin/teams:
    get:
      tags:
        - Team
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    post:
      tags:
        - Team
      summary: create team
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Team'
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/teams/{id}:
    get:
      tags:
        - Team
      summary: read team
      description: Returns a single typeshare
      parameters:
        - name: id
          in: path
          description: ID of Team to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get data success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - Team
      summary: Update Team
      parameters:
        - name: id
          in: path
          description: id that need to be update
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_Team'
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - Team
      summary: Deletes a Team
      description: delete a Team
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get data success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
          example: sgod888@gmail.com
        password:
          type: string
          example: Sgod123@
        full_name:
          type: object
          properties:
            last:
              type: string
              example: sgod
            first:
              type: string
              example: team
        birthday:
          type: object
          properties:
            day:
              type: number
              example: 10
            month:
              type: number
              example: 11
            year:
              type: number
              example: 2021
        code:
          type: string
          example: 770182
        sex:
          type: string
          enum: ['Male', 'Female', 'Other']
          example: Male
        address:
          type: string
          example: 400 Dang Dinh Tuong, Phuong An Khanh, Q2
...
    Team:
      type: object
      properties:
        avatar:
          type: string
          format: binary
        full_name:
          type: object
          properties:
            first:
              type: string
              example: Sgod
            last:
              type: string
              example: 5904
        email:
          type: string
          example: infor@gmail.com
        phone:
          type: string
          example: 0343242992
        birthday:
          type: object
          properties:
            day:
              type: number
              example: 4
            month:
              type: number
              example: 6
            year:
              type: number
              example: 2001
        sex:
          type: string
          enum:
            - Male
            - Female
            - Other
          example: Male
        academicLevel:
          type: string
          example: 'Thac Si'
        experience:
          type: string
          example: 'Sgod'
        position:
          type: string
          example: 6438f2938946235f75365baa
        groupId:
          type: string
          example: 6438f2938946235f75365baa
    _Team:
      type: object
      properties:
        avatar:
          type: string
          format: binary
        full_name:
          type: object
          properties:
            first:
              type: string
              example: Sgod
            last:
              type: string
              example: 5904
        email:
          type: string
          example: infor@gmail.com
        phone:
          type: string
          example: 0343242992
        birthday:
          type: object
          properties:
            day:
              type: number
              example: 4
            month:
              type: number
              example: 6
            year:
              type: number
              example: 2001
        sex:
          type: string
          enum:
            - Male
            - Female
            - Other
          example: Male
        academicLevel:
          type: string
          example: 'Thac Si'
        experience:
          type: string
          example: 'Sgod'
    Account:
      type: object
      required:
        - full_name
        - username
        - sex
        - address
      properties:
        avatar:
          type: string
          format: binary
        full_name:
          type: object
          properties:
            first:
              type: string
              example: Sai
            last:
              type: string
              example: Odessa
        username:
          type: string
          example: sgodCompany
        birthday:
          type: object
          properties:
            day:
              type: integer
              example: 29
            month:
              type: integer
              example: 9
            year:
              type: integer
              example: 1997
        sex:
          type: string
          enum:
            - Male
            - Female
            - Other
          default: Male
        address:
          type: string
          example: 2 Street No. 2 Tan Truong Residential Area, Phu Thuan, District 7, Ho Chi Minh City, Vietnam
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer